+++ Project1/src/tmp102_sensor.c
@@ -10,7 +10,7 @@
 #include "tmp102_sensor.h"
 
 
-int TMP102_getTemp(double *temp, TEMPERATURE_UNIT_T unit)
+int TMP102_getTemp(float *temp, TEMPERATURE_UNIT_T unit)
 {
     uint8_t buff[2] = {0};
     int ret = I2Cmaster_read_bytes(TMP102_SLAVE_ADDR, TMP102_REG_TEMPERATURE, buff, sizeof(buff));
@@ -18,11 +18,11 @@ int TMP102_getTemp(double *temp, TEMPERATURE_UNIT_T unit)
     if(temp_raw & 0x800)
     {
         temp_raw = (~temp_raw) + 1; 
-        *temp = (-1) * (double)temp_raw * 0.0625;
+        *temp = (-1) * (float)temp_raw * 0.0625;
     }
     else
     {
-        *temp = ((double)temp_raw)*0.0625;
+        *temp = ((float)temp_raw)*0.0625;
     }
 
     if(unit == FAHREN)
